requirements.txt
python-telegram-bot==13.15
requests==2.31.0
python-dotenv==1.0.0
APScheduler==3.6.3
psycopg2-binary==2.9.9
redis==5.0.1
pytz==2024.1